// src/types.ts
export type AuthMethod = 'cookies' | 'credentials' | 'api';

export interface AuthConfig {
  method: AuthMethod;
  data: CookieAuth | CredentialsAuth | ApiAuth;
}

export interface CookieAuth {
  cookies: string[];
}

export interface CredentialsAuth {
  username: string;
  password: string;
  email?: string;
  twoFactorSecret?: string;
}

export interface ApiAuth {
  apiKey: string;
  apiSecretKey: string;
  accessToken: string;
  accessTokenSecret: string;
}
// ...
export class TwitterMcpError extends Error {
  constructor(
    message: string,
    public readonly code: string,
    public readonly status?: number
  ) {
    super(message);
    this.name = 'TwitterMcpError';
  }
}